code/requirements.txt
antlr4-python3-runtime==4.9.3
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
braceexpand==0.1.7
brotlipy==0.7.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet @ file:///tmp/build/80754af9/chardet_1607706768982/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.13.0
conda-build==3.21.9
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1649105789509/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652083456434/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
decord==0.6.0
filelock @ file:///opt/conda/conda-bld/filelock_1647002191454/work
fsspec==2023.1.0
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
huggingface-hub==0.14.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==6.6.0
iopath==0.1.10
ipython @ file:///tmp/abs_94gruux8u8/croots/recipe/ipython_1659529858706/work
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2==2.10.1
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1651563629415/work
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
pandas==1.3.5
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
pkginfo @ file:///tmp/build/80754af9/pkginfo_1643162084911/work
portalocker==2.7.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
psutil @ file:///tmp/build/80754af9/psutil_1612298016854/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyYAML==6.0
regex==2023.5.5
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016701961/work
sentencepiece==0.1.98
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
timm==0.6.13
tokenizers==0.13.3
torch==1.12.1
torchtext==0.13.1
torchvision==0.13.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
transformers==4.28.0
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webdataset==0.2.48
zipp==3.15.0
